update Needlefish, update PacketAccepted event
- PacketAccepted now correctly sends the deserialized packet instead of just the header
- for consistency PacketManager now uses int for id instead of casting to ushort

+++ Swordfish.Library/Networking/PacketManager.cs
@@ -91,7 +91,7 @@ namespace Swordfish.Library.Networking
 
                     if (typeof(IDataBody).IsAssignableFrom(type))
                     {
-                        ushort id = (ushort)(packetAttribute.PacketID ?? type.FullName.ToSeed());
+                        int id = packetAttribute.PacketID ?? type.FullName.ToSeed();
                         PacketDefinition definition = new PacketDefinition {
                             ID = id,
                             Type = type,
